Extract AddAlias and TransferGrain ledger rows into their own components
